refactor: 폴더명 컴포넌트명 일치
