+++ src/containers/random-pick/playground/card/card.tsx
