+++ src/containers/random-pick/playground/result-card/result-card.tsx
